+++ space_battle/src/server/game_elements/ProjectileGoingUp.java
@@ -2,7 +2,7 @@ package server.game_elements;
 
 public class ProjectileGoingUp extends Projectile {
 	
-	ProjectileGoingUp(int x, int y) {
+	ProjectileGoingUp(double x, double y) {
 		super(x,y);
 	}
 	
@@ -14,13 +14,13 @@ public class ProjectileGoingUp extends Projectile {
 	@Override
 	public boolean isHit(GameElement object){
 		// projectile-related
-		int top = getCoordY() - super.getProjectileheight()/2;
-		int projX = getCoordX();
+		int top = (int) (getCoordY() - super.getProjectileheight()/2);
+		int projX = (int) getCoordX();
 		// object-related
 		int objWidth = object.getHitBox().getWidth();
 		int objHeight = object.getHitBox().getHeigth();
-		int objX = object.getCoordX();
-		int objY = object.getCoordY();
+		int objX = (int) object.getCoordX();
+		int objY = (int) object.getCoordY();
 		
 		
 		if( (projX > objX - objWidth/2) && (projX < objX + objWidth/2) ){ // projectile's coordX in range
