+++ space_battle/src/server/game_elements/ProjectileLaser.java
@@ -4,7 +4,7 @@ public class ProjectileLaser extends ProjectileGoingUp {
 	
 	// identical to ProjectileGoingUp, but goes through Hostiles, exploding them at first hit
 	
-	ProjectileLaser(int x, int y, Player whoShotThis) {
+	ProjectileLaser(double x, double y, Player whoShotThis) {
 		super(x,y);
 	}
